重进存档会击杀和开箱卡

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/MossyClump.java
@@ -26,9 +26,9 @@ public class MossyClump extends Trinket {
 	@Override
 	public String statsDesc() {
 		if (已鉴定()){
-			return Messages.get(this, "stats_desc", (int)(100*overrideNormalLevelChance(等级())));
+			return Messages.get(this, "stats_desc", 100*overrideNormalLevelChance(等级()));
 		} else {
-			return Messages.get(this, "stats_desc", (int)(100*overrideNormalLevelChance(0)));
+			return Messages.get(this, "stats_desc", 100*overrideNormalLevelChance(0));
 		}
 	}
